pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-anywhere"
version = "0.8.0"
description = "Unified gateway for Model Context Protocol servers"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "MCP Anywhere Team", email = "team@mcpeverywhere.com"}
]
dependencies = [
    "Flask>=2.0",
    "Flask-SQLAlchemy>=2.5",
    "Flask-WTF>=1.0",
    "Flask-Login>=0.6.0",
    "flask-sock>=0.7.0",
    "flask-cors>=4.0.0",
    "bcrypt>=4.0.0",
    "wtforms>=3.1.0",
    "fastmcp>=2.9.0",
    "llm-sandbox>=0.2.0",
    "mcp>=1.1.2",
    "anthropic",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.0",
    "docker",
    "loguru>=0.7.0",
    "pydantic>=2.10.3",
    "PyJWT>=2.8.0",
    "tenacity>=8.2.0",
    "uvicorn[standard]>=0.24.0",
    "starlette>=0.36.0",
    "a2wsgi>=1.10.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "pytest-asyncio>=0.20",
    "pytest-flask>=1.2.0",
    "pytest-mock>=3.10.0",
    "respx>=0.20.0"
]

[project.scripts]
mcp-anywhere = "mcp_anywhere.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mcp_anywhere = ["templates/**/*.html", "static/**/*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true 